update compute shader

+++ examples/9.self/compute_shader/compute_image1d/main.cpp
@@ -113,7 +113,7 @@ void DoCompute()
 
     for (int i = 0; i < kArraySize; i++)
     {
-        inputData[i] = i;
+        inputData[i] = (float)i;
     }
 
     glBindTexture(GL_TEXTURE_1D, inputTexId);
